- separated parallel_for_internals.hpp and parallel_for.hpp - added the possibility to control the grain size in the static scheduling   policy of the ParallelFor(Reduce) pattern(s) - ff_node::eosnotify interface changed:    virtual void eosnotify(int id=-1) --> virtual void eosnotify(ssize_t id=-1)

+++ tests/test_masterworker.cpp
@@ -76,7 +76,7 @@ public:
         if (--numtasks == 0 && eosreceived) return NULL;
         return GO_ON;	
     }
-    void eosnotify(int id) {
+    void eosnotify(ssize_t id) {
         if (id == -1)  {
             eosreceived = true;
             if (numtasks == 0) {
